# services/control-plane/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RunIt API
  description: |
    API for managing projects, runs, secrets, and deployments in RunIt.

    ## Authentication
    All endpoints (except health checks and metrics) require a valid Bearer token in the Authorization header.
    Use an API key or JWT token.

    ## Rate Limiting
    - Authenticated users: 120 requests per minute
    - Anonymous users: 60 requests per minute
  version: 0.1.0
  contact:
    name: RunIt
    url: https://runit.dev
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.runit.dev
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  schemas:
...
    Endpoint:
      type: object
      required:
        - id
        - method
        - path
      properties:
        id:
          type: string
          description: Endpoint identifier (path-based)
          example: "GET_/users"
        endpoint_id:
          type: string
          description: Endpoint identifier (alias for id)
          example: "GET_/users"
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
          description: HTTP method
          example: "GET"
        path:
          type: string
          description: URL path pattern
          example: "/users/{id}"
        summary:
          type: string
          description: Brief description of the endpoint
          example: "Get user by ID"
        description:
          type: string
          description: Detailed description of the endpoint
    CreateProjectRequest:
      type: object
      required:
        - name
        - source_type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Project display name
          example: "My API Project"
        source_type:
          type: string
          enum: [zip, github]
          description: Source type for the project code
          example: "zip"
        zip_data:
          type: string
          format: byte
          description: Base64-encoded ZIP file (required if source_type is 'zip')
        github_url:
          type: string
          format: uri
          pattern: "^https://github\\.com/[\\w\\-\\.]+/[\\w\\-\\.]+(?:\\.git)?$"
          description: GitHub repository URL (required if source_type is 'github')
          example: "https://github.com/user/repo"
        github_ref:
          type: string
          description: Git branch or tag to clone
          example: "main"
    CreateProjectResponse:
      type: object
      required:
        - project_id
        - project_slug
        - version_id
        - version_hash
        - status
      properties:
        project_id:
          type: string
          format: uuid
          description: Created project ID
        project_slug:
          type: string
          description: URL-friendly project slug
        version_id:
          type: string
          format: uuid
          description: Created version ID
        version_hash:
          type: string
          description: Version hash
        status:
          type: string
          enum: [draft]
          description: Initial project status
        detected_env_vars:
          type: array
          items:
            type: string
          description: Environment variables detected in the code
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/Endpoint'
          description: Extracted API endpoints
...
    ListProjectsResponse:
      type: object
      required:
        - projects
        - total
      properties:
        projects:
          type: array
          items:
            type: object
            properties:
              project_id:
                type: string
                format: uuid
              project_slug:
                type: string
              name:
                type: string
              status:
                type: string
                enum: [draft, deploying, live, failed]
              latest_version:
                type: string
              created_at:
                type: string
                format: date-time
              updated_at:
                type: string
                format: date-time
        total:
          type: integer
          description: Total number of projects

    # ==================== Runs ====================
...
    SubscriptionResponse:
      type: object
      required:
        - tier
        - status
        - usage
        - limits
      properties:
        tier:
          type: string
          enum: [free, pro, team]
          description: Current subscription tier
        status:
          type: string
          enum: [active, past_due, canceled]
          description: Subscription status
        current_period_end:
          type: string
          format: date-time
          nullable: true
        usage:
          type: object
          properties:
            cpu_runs:
              type: integer
            gpu_runs:
              type: integer
            projects_count:
              type: integer
        limits:
          type: object
          description: Tier-specific resource limits
    CheckoutRequest:
      type: object
      required:
        - tier
      properties:
        tier:
          type: string
          enum: [pro, team]
          description: Target subscription tier
    CheckoutResponse:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: Stripe Checkout session URL
    PortalResponse:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: Stripe Customer Portal session URL

    # ==================== Templates ====================
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Project not found"
        code:
          type: string
          description: Error code for programmatic handling
          example: "PROJECT_NOT_FOUND"
        details:
          type: string
          description: Additional error details
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Validation error message
          example: "Missing required fields: name, source_type"
        received:
          type: string
          description: Value that was received (for debugging)
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          examples:
            missingFields:
              value:
                error: "Missing required fields: name, source_type"
            invalidFormat:
              value:
                error: "Invalid secret key format. Use UPPERCASE_SNAKE_CASE (A-Z, 0-9, _)"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per minute
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in the current window
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
paths:
...
  /v1/projects:
    post:
      tags: [Projects]
      summary: Create a new project
      description: |
        Creates a new project from either a ZIP file upload or a GitHub repository.
        The code is analyzed to extract OpenAPI endpoints automatically.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            examples:
              zipUpload:
                summary: ZIP file upload
                value:
                  name: "My API"
                  source_type: "zip"
                  zip_data: "UEsDBBQAAAAIAA..."
              githubImport:
                summary: GitHub import
                value:
                  name: "My GitHub API"
                  source_type: "github"
                  github_url: "https://github.com/user/repo"
                  github_ref: "main"
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Project limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [Projects]
      summary: List all projects
      description: Returns all projects owned by the authenticated user
      operationId: listProjects
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/billing/subscription:
    get:
      tags: [Billing]
      summary: Get subscription and usage
      description: Returns current subscription tier, status, usage counters, and tier limits
      operationId: getSubscription
      responses:
        '200':
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/billing/checkout:
    post:
      tags: [Billing]
      summary: Create checkout session
      description: Creates a Stripe Checkout session for upgrading to a paid tier
      operationId: createCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          description: Billing not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/billing/portal:
    post:
      tags: [Billing]
      summary: Create portal session
      description: Creates a Stripe Customer Portal session for managing subscription
      operationId: createPortalSession
      responses:
        '200':
          description: Portal session URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No billing account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          description: Billing not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/billing/webhook:
    post:
      tags: [Billing]
      summary: Stripe webhook handler
      description: |
        Handles Stripe webhook events for subscription lifecycle management.
        Excluded from auth middleware; verified by Stripe signature.
      operationId: stripeWebhook
      security: []
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema:
            type: string
          description: Stripe webhook signature
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        '400':
          description: Invalid signature or missing header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Billing or webhook not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== Templates ====================
...